# feat(perf): Memoize component scope resolution lexing and parsing