Many components can include the same files via `<script/>` tag, and we previously lexed, preprocessed, and parsed those files every time they were encountered in any component.  Even if a component and its extension both import `uri="pkg:/components/foo.brs"`, we'd parse it twice.  Bummer!

We can save a significant amount of time by memoizing that lexing, preprocessing, and parsing process across components.