@sjbarag 2020-07-27
As far as real-world numbers go, I ran a simple "hello world" script from the root of a large production channel (300+ components), which triggered a full lex/parse cycle across all components.  With some timing logs added to `src/index.ts`, here's the results:


#### Before
```
$ /path/to/brs/bin/cli.js foo.brs
manifest parsing: 3.225ms
component definitions: 194.009ms
script path rewriting: 7.750ms
create lexerParserFn: 0.115ms
create interpreter: 11839.352ms
lex & parse explicit files: 0.838ms
pre-interpreter steps: 12048.050ms
hello, world
```

#### After
```
$ /path/to/brs/bin/cli.js foo.brs
manifest parsing: 2.883ms
component definitions: 194.693ms
script path rewriting: 8.228ms
create lexerParserFn: 0.078ms
create interpreter: 1145.850ms
lex & parse explicit files: 4.493ms
pre-interpreter steps: 1358.864ms
hello, world
```